refactor: Move into subdirectories
