+++ src/components/Modules/components/moduleUpdate/UpdateModuleButton/UpdateModuleButton.test.tsx
@@ -28,11 +28,11 @@ vi.mock('resources/Namespaces/YamlUpload/useUploadResources', () => ({
   useUploadResources: () => uploadResourcesMock,
 }));
 
-vi.mock('../../community/communityModulesHelpers', () => ({
+vi.mock('../../../community/communityModulesHelpers', () => ({
   fetchResourcesToApply: vi.fn().mockResolvedValue(undefined),
 }));
 
-vi.mock('./DeleteOldModulesCheck', async () => {
+vi.mock('../DeleteOldModulesCheck/DeleteOldModulesCheck', async () => {
   const { useState } = await import('react');
   return {
     useDeleteOldModuleTemplates: (_tpls: any[]) => {
